refactor: extract handleCancel from tools.ts to cancel.ts
