+++ plugins/exarchos/servers/exarchos-mcp/src/__tests__/workflow/index.test.ts
@@ -38,6 +38,11 @@ vi.mock('../../workflow/next-action.js', () => ({
   handleNextAction: vi.fn().mockResolvedValue({ success: true, data: { action: 'DONE' } }),
 }));
 
+// Mock cancel handler (now in its own module)
+vi.mock('../../workflow/cancel.js', () => ({
+  handleCancel: vi.fn().mockResolvedValue({ success: true, data: {} }),
+}));
+
 // Mock all tool handlers so we don't need real state files
 vi.mock('../../workflow/tools.js', () => ({
   handleInit: vi.fn().mockResolvedValue({ success: true, data: { phase: 'ideate' } }),
@@ -62,10 +67,10 @@ import {
   handleSummary,
   handleReconcile,
   handleTransitions,
-  handleCancel,
   handleCheckpoint,
 } from '../../workflow/tools.js';
 import { handleNextAction } from '../../workflow/next-action.js';
+import { handleCancel } from '../../workflow/cancel.js';
 
 // ─── Tests ───────────────────────────────────────────────────────────────────
 
@@ -370,6 +375,9 @@ describe('MCP Server Entry Point', () => {
         }));
         vi.doMock('../../workflow/next-action.js', () => ({
           handleNextAction: vi.fn(),
+        }));
+        vi.doMock('../../workflow/cancel.js', () => ({
+          handleCancel: vi.fn(),
         }));
         vi.doMock('../../stack/tools.js', () => ({
           handleStackStatus: vi.fn(),
